requirements.txt
django<1.9
django-suit
django-nested-inline
django-crispy-forms